store-weather-forecast/requirements.txt
google-cloud-firestore
requests
Flask
